fix(db): workflow_executions.total_time_ms 类型对齐 BIGINT (i32→i64)
DDL 中 total_time_ms 为 BIGINT（v100_consolidated.rs），Rust 之前误用 i32，
会导致 PG 上 INT8→INT4 解码失败。全链路统一改为 i64：
- entities Model 字段
- harness WorkflowExecutionData / WorkflowExecutionRepository trait
- dao repo 实现与 DaoWorkflowExecutionRepository
- rt-workflow 引擎调用处（as i64）
- commands ExecutionSummaryResponse DTO
- 测试 mock / test_support 同步

+++ src-tauri/crates/dao/src/repo/workflow_execution.rs
@@ -43,7 +43,7 @@ pub async fn update_workflow_execution_status(
     status: &str,
     output_result: Option<&str>,
     node_executions: Option<&str>,
-    total_time_ms: Option<i32>,
+    total_time_ms: Option<i64>,
 ) -> Result<bool> {
     let execution = workflow_executions::Entity::find_by_id(id).one(db).await?;
     if let Some(e) = execution {
