+++ src-tauri/crates/rt-workflow/tests/per_node_exec_ctx_vars.rs
@@ -54,7 +54,7 @@ impl WorkflowExecutionRepository for MockWorkflowExecRepo {
         _status: &str,
         _output_result: Option<&str>,
         _node_executions: Option<&str>,
-        _total_time_ms: Option<i32>,
+        _total_time_ms: Option<i64>,
     ) -> Result<bool, String> {
         Ok(true)
     }
